Fix wrong layer bug #10
https://github.com/cookbrite/flyingcloud/issues/10
aka flyingcloud.base - WARNING - TestRunner build does nothing
for building any layer.

This was due to a bug in [argparse](https://bugs.python.org/issue9351)'s
handling of `set_defaults` in subparsers that was fixed in
[Python 2.7.10](https://www.python.org/downloads/release/python-2710/).

The FlyingCloud fix is to not set a default value for `layer_inst`
in the parent parser. It was unnecessary.